# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for Objective Zero"
  version: "1.0.0"
  title: "Objective Zero API"
  contact:
    email: "you@your-company.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "107.170.82.21:3000"
basePath: "/mobile"
...
schemes:
- "http"
paths:
...
  /facebook:
    post:
      summary: "User Sign In"
      description: "facebook regsitartion"
      operationId: "facebook"
      produces:
      - "application/json"
      parameters:
        - in: body
          name: facebook Ragistration
          description: The user to create.
          schema:
            type: object
            required:
              - email
            properties:
              email:
                type: string
              fb_id:
                type: string
              deviceId:
                type: string
              deviceType:
                type: string 
              name:
                type: string   
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "user"
  /signout:
    post:
      summary: "User Sign Out"
      description: "Sign Out User"
      operationId: "facebook"
      produces:
      - "application/json"
      parameters:
        - name: access_token
          in: header
          description: Access Token
          required: true
          type: string
          format: string
        - in: body
          name: facebook Ragistration
          description: The user to create.
          schema:
            type: object
            required:
              - email
            properties:
              email:
                type: string 
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "user"
...